+++ packages/firebase_storage_util/CHANGELOG.md
@@ -3,8 +3,9 @@
 # Changelog 
 All notable changes to this project will be documented in this file.
 
-<!-- ## [Unreleased] -->
-
+## [Unreleased]
+### Added
+- `deleteFromPath()`, `deleteFromUrl()`, `delete()` methods. 
 
 
 ## [0.0.1] - 2021-07-10